+++ components/Footer.tsx
@@ -1,0 +1,25 @@
+export default function () {
+  return (
+    <footer class="mt-8 col-full">
+      <div class="card">
+        <ul>
+          <li>
+            📡 <a href="rss.xml">RSS</a>
+          </li>
+          <li>
+            🐙 <a href="https://github.com/CarcajadaArtificial">GitHub</a>
+          </li>
+          <li>
+            🐘 <a href="https://techhub.social/@carcajada">Mastodon</a>
+          </li>
+          <li>
+            🗞️{" "}
+            <a href="https://app.follow.is/share/feeds/137022670373136384">
+              Folo
+            </a>
+          </li>
+        </ul>
+      </div>
+    </footer>
+  );
+}